code/loan_work/src/requirements.txt
langgraph
langchain_openai
langchain_core
langchain_voyageai
pandas
faiss-cpu
python-dotenv==1.0.0
langchain-community
pydantic
pydantic_core
langchain_groq

